Use a layout_height of 0dip instead of wrap_content for better performance
When only a single widget in a LinearLayout defines a weight, it is more
efficient to assign a width/height of 0dp to it since it will absorb all
the remaining space anyway. With a declared width/height of 0dp it does
not have to measure its own size first.